requirements.txt
aquasec-api==0.0.2
click==8.1.3
